modification

+++ kernel/mgmt/process_mgmt.c
@@ -6,7 +6,7 @@ struct Process_Control_Block* get_pcb_head();
 struct Process_Control_Block{
 	uint32_t pid;
 	uint32_t pstat;
-	uint8_t apic_id;
+	uint32_t apic_id;
 	uint32_t text_start;
 	uint32_t text_size;
 	uint32_t bss_start;
